project/requirements.txt
asgiref==3.2.10
astroid==2.3.3
autopep8==1.5
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0.8
django-3-jet==1.0.8
django-admin-sortable2==0.7.6
django-ckeditor==5.9.0
django-colorfield==0.3.2
django-debug-toolbar==3.1.1
django-extensions==3.0.4
django-js-asset==1.2.2
django-mptt==0.11.0
django-redis==4.12.1
django-robots==4.0
django-simple-history==2.11.0
flake8==3.8.3
idna==2.10
install==1.3.4
isort==4.3.21
komtet-kassa-sdk==5.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.4
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.1
pylint==2.4.4
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
wrapt==1.11.2
XlsxWriter==1.2.7
